# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar-IndigoPay API
  version: 1.0.0
  description: |
    OpenAPI 3.0 documentation for the backend REST API used by Stellar-IndigoPay.
    The API exposes project discovery, donations, profiles, leaderboard, notifications,
    ratings, updates, jobs, stats, and admin management endpoints.
  contact:
    name: Stellar-IndigoPay Team
    url: https://github.com/Stellar-IndigoPay/Stellar-IndigoPay
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server  
  - url: https://api.stellarindigopay.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PublicKeyParam:
      name: publicKey
      in: path
      required: true
      description: Stellar public key.
      schema:
        type: string
        pattern: "^G[A-Z0-9]{55}$"
...
  responses:
    BadRequest:
      description: The request payload or query parameters are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimitExceeded:
      description: The request has been rate-limited. All endpoints carry rate-limit response headers (X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset) and return 429 with a Retry-After header when the limit is exceeded. See the API documentation for per-endpoint tiers.
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying.
          schema:
            type: integer
        X-RateLimit-Limit:
          description: Maximum requests allowed (or bucket capacity for token-bucket endpoints).
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in the current window (or tokens left in the bucket).
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Epoch seconds when the limit resets (or next token becomes available).
          schema:
            type: integer
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Too many requests — Try again later.
              retryAfter:
                type: integer
                description: Epoch seconds after which a retry may succeed.
                example: 1742169602
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid request payload
      required:
        - error
...
    Profile:
      type: object
      properties:
        publicKey:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        totalDonatedXLM:
          type: string
        projectsSupported:
          type: integer
        badges:
          type: array
          items:
            type: object
            properties:
              tier:
                type: string
              earnedAt:
                type: string
                format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateProfileRequest:
      type: object
      required: [publicKey]
      properties:
        publicKey:
          type: string
          pattern: "^G[A-Z0-9]{55}$"
        displayName:
          type: string
          maxLength: 30
        bio:
          type: string
          maxLength: 300
...
paths:
...
  /api/profiles/{publicKey}:
    get:
      tags: [Profiles]
      summary: Get a profile by public key
      operationId: getProfileByPublicKey
      description: Retrieve a donor profile by their Stellar public key, including display name, bio, badges, and donation stats.
      parameters:
        - $ref: "#/components/parameters/PublicKeyParam"
      responses:
        "200":
          description: Profile data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
  /api/profiles:
    post:
      tags: [Profiles]
      summary: Create or update a profile
      operationId: upsertProfile
      description: Create a new donor profile or update an existing one with display name and bio.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileRequest"
      responses:
        "200":
          description: Profile upserted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Profile"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
...